GRID-4067 Move Value tests to seperate Test Suite
